refactor(stories): consolidate Misc group into misc.stories.ts
FullMesh, MeshWithHoles, and OnDemandRendering move from the removed
experiments.stories.ts (which already registered under Examples/Misc)
into misc.stories.ts next to RepulsionBenchmark; impls move to
src/stories/misc/. One file per sidebar group again.
