+++ src/stories/misc/full-mesh.ts
