+++ src/stories/misc/mesh-with-holes.ts
